requirements.txt
absl-py==2.3.1
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.50.0
Deprecated==1.2.18
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.19.1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.1
grpcio==1.75.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.11.3
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.3
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
opencv-python==4.11.0.86
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.1.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shap==0.48.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
